Milestone changes (TIP-29) (#1347)
* Milestone timestamp as u32

* Add milestone options

* Fmt

* MilestoneOption kind as u8

* Add new milestone fields

* Change Milestone::id

* Add PowMilestoneOption

* Address comment

* Update .gitignore

* Fix snapshot parsing

* Fix MilestoneOptionCount size

* Adapt WTF to new merkle roots

* Check if previous milestone is in the past cone

* Fix tests

* Use != instead of !.eq

* Remove enum comments

* Update bee-message/src/payload/milestone/option/mod.rs


* Map instead of fold

* Remove println

+++ bee-api/bee-rest-api/src/types/responses.rs
@@ -31,7 +31,7 @@ pub struct StatusResponse {
     #[serde(rename = "isHealthy")]
     pub is_healthy: bool,
     #[serde(rename = "latestMilestoneTimestamp")]
-    pub latest_milestone_timestamp: u64,
+    pub latest_milestone_timestamp: u32,
     #[serde(rename = "latestMilestoneIndex")]
     pub latest_milestone_index: u32,
     #[serde(rename = "confirmedMilestoneIndex")]
